feat: extract scene class
